update card mongoose model to match new card
